Changed comments language to English
This commit updates all the comments in the code to English. Previously, the comments were in a different language, which made it difficult for non-native speakers to understand the code. By changing the comments to English, we hope to make the code more accessible to a wider audience and improve its readability. No functional changes have been made to the code itself.